+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Add freetext search field.
 - Fix listing search custom form prefix.
+- Fix collection tiles listing retrieval to include configuration of the target collection.
 
 
 0.1.1 (2014-07-15)